+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Strict option for loading Modulus checkpoints.
 - Regression only or diffusion only inference for CorrDiff.
 - Support for organization level model files on NGC file system
+- Augmentation pipeline for CorrDiff.
 - Code logging for CorrDiff via wandb
 
 ### Changed